mavis/blowfish.h: sizeof(a_char) should be sizeof(u_long), not sizeof(u_int). This fixes https://github.com/MarcJHuber/event-driven-servers/security/advisories/GHSA-vc85-v569-2r9f

+++ mavis/blowfish.h
@@ -12,7 +12,7 @@
 
 typedef union {
     char s[1];
-    u_int n[1];
+    u_long n[1];
 } a_char;			/* aligned char */
 
 struct blowfish_ctx;
